requirements.txt
pyspark==3.3.2
delta-spark==2.2.0
pytest==7.4.0
duckdb==0.9.1
python-dateutil==2.8.2
great-expectations==1.1.0
